backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.1  # passlib 1.7.4 ile uyumlu son sürüm; 4.1+ "72 bytes" check ile patlıyor
python-multipart==0.0.20
email-validator==2.2.0
